add test for zs3server in separate files
