+++ tests/api_tests/list_objects_test.go
@@ -1,0 +1,24 @@
+package api_tests
+
+import (
+	"testing"
+
+	"github.com/0chain/system_test/internal/api/util/test"
+	"github.com/stretchr/testify/require"
+)
+
+func TestListObjects(testSetup *testing.T) {
+	t := test.NewSystemTest(testSetup)
+	t.Parallel()
+	t.Run("ListObjects should return 200 when all the parameters are correct", func(t *test.SystemTest) {
+		queryParams := map[string]string{
+			"accessKey":       AccessKey,
+			"secretAccessKey": SecretAccessKey,
+			"action":          "listObjects",
+			"bucketName":      "root",
+		}
+		resp, err := zs3Client.BucketOperation(t, queryParams, map[string]string{})
+		require.Nil(t, err)
+		require.Equal(t, 200, resp.StatusCode())
+	})
+}
